requirements.txt
numpy
pandas
scipy
seaborn
hdbscan
alive-progress
lxml
html5lib
dask
dask[dataframe]
selenium
dash
matplotlib
scikit-learn
umap-learn
plotly
streamlit
vnstock3
pyfolio-reloaded
empyrical-reloaded
PyPortfolioOpt
jupyter